remove beta note
@CryZe and others did enough testing over the last few weeks to
make us feel comfortable that the API can deliver what people
expect.

There will most likely be quite a lot of additions to follow (like
specifying multiple players when submitting a run), but those
should not affect the grand scheme of things. Same goes for the
current rate limit, which could be increased in the next few months
depending on how the usage develops.